requirements.txt
opencv-python
pillow
pytesseract
requests
fastapi
uvicorn
python-dotenv
requests
pyspellchecker